+++ erpnext/app/greenleaf_website/greenleaf_website/green_leaf_website/__init__.py
@@ -1,0 +1,1 @@
+# Green Leaf Website module.